requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==3.1.12
djongo==1.3.7
dnspython==2.7.0
gunicorn==23.0.0
mongoengine==0.29.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.11
pymongo==3.11.4
python-dotenv==1.2.1
pytz==2025.2
sqlparse==0.2.4
typing_extensions==4.15.0
whitenoise==6.6.0
